update 7
I have worked a little with the media queries after getting tons of help. still something up with it, så the landing page doesnt look good anymore. On the other side, i hva successfully completed the "about" page. i have laso today kept on working with the dektop version of the assignment. my plan for tomorrow is to finish the desktop version, and then get back to the media queries and rewrite the CSS to follow DRY.