Update 9
I have completed the desktop version of what i made in Figma. Now i am working on the scaling (media queries), and i have completed the 3 of the pages already. another problem that has occured, is the placement of the footer, that is attaching itself to the eloement above. i will for now focus on the scaling, then fix the footer and nav at last, since when i do them, it will appear on all pages. if i have time after that, i will add a login form, even if it wasnt planned yet.