Continue rewriting the CSS
I have completed the rewrite of the CSS that is connected to the index page. The rewrite for footer and navigationbar will be common for all pages. I will add those to the other pages later. next step is to slowly rewrite the CSS for the other pages as well. This time all the CSS will be in one file instead of different files. i will at least try. Checking another page tomorrow.